Black box fix for disguised Spy and Ubered enemy (#9)
* Black box fix for disguised Spy and Ubered enemy

Added if checks for when a Spy is disguised and when an enemy is Ubered to not heal the attacker

"The Soldier will not receive health if the target is a disguised Spy or an enemy under the effects of an ÜberCharge; however, health will still be awarded if the target is a Scout under the effects of Bonk! Atomic Punch. "

https://wiki.teamfortress.com/w/index.php?title=Black_Box&oldid=1922390

Note: only tested on bots using addcond commands in itemtest - if checks not tested yet if it works on rockets reflected by Pyros.

* formatting

---------

Co-authored-by: rsedxcftvgyhbujnkiqwe <102834306+rsedxcftvgyhbujnkiqwe@users.noreply.github.com>