Sandman ball recharge time historical fix (#10)
Turns out the pre-Inferno Sandman ball recharge rate used to be 15 seconds. current recharge time is 10 seconds.

I just added a +50% recharge time for the ball using one attribute. Hopefully nothing breaks. Credits to huutti for pointing this out.

https://wiki.teamfortress.com/w/index.php?title=Sandman&oldid=2227698
"A baseball that does not hit an enemy will remain on the ground and disappear after 15 seconds, effectively the time it takes the Scout to gain a new one."