(Fix) Enforcer damage bonus on buildings (#80)
Uses attributes instead of an involved revert to mimic the old Enforcer. 

The 20% damage bonus attribute is used so that both players and buildings get affected by the bonus damage, while the -16.667% damage bonus while disguised cancels out the 20% damage bonus so that it acts like it has no damage bonus when the Spy fires when disguised.