(Fix) Prevent Reverted DR from losing 70% cloak when hit by Pomson at close range (#175)
* dead ringer + pomson bug fix

tested, seems to work.
hopefully there are no issues

* add revert DR check

* bug fix for DR losing 70% cloak when hit by pomson at close range

also includes debugging lines

* Fix for unreverted Dead Ringer and reverted Pomson

Note that if Dead Ringer is reverted and Pomson is NOT reverted, the reverted Dead Ringer will lose 70% cloak on hit

* comment debugging lines

* add check for vanilla pomson and revert DR

i have yet to figure out how to make the uber/cloak drain only happen once for the release variant of the pomson.
anyhow, the default variant of the pomson should be more accurate now.